Fix routing module for entire project
Remove wrong routing modules
Add correct routing module
Route paths fix
