--- src/app/app-router/app-router.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { AppRouterRoutingModule } from './app-router-routing.module';
-
-
-@NgModule({
-  declarations: [],
-  imports: [
-    CommonModule,
-    AppRouterRoutingModule
-  ]
-})
-export class AppRouterModule { }
