+++ src/app/app.component.html
@@ -1,2 +1,1 @@
-<app-artist-details></app-artist-details>
-<app-artists></app-artists>
+<router-outlet></router-outlet>
